+++ upcloud/provider.go
@@ -69,7 +69,7 @@ func Provider() *schema.Provider {
 				Type:        schema.TypeInt,
 				Optional:    true,
 				Default:     120,
-				Description: "The duration (in seconds) that the provider waits for a HTTP request to towards UpCloud API to complete. Defaults to 120 seconds",
+				Description: "The duration (in seconds) that the provider waits for an HTTP request towards UpCloud API to complete. Defaults to 120 seconds",
 			},
 		},