+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [3.2.0] - 2023-12-19
+
 ### Added
 
 - load_balancer: `upcloud_loadbalancer_backend_tls_config` resource for backend TLS config management
@@ -428,7 +430,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.1.1...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.1.1...v3.2.0
 [3.1.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.1.0...v3.1.1
 [3.1.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.0.3...v3.1.0
 [3.0.3]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.0.2...v3.0.3